Clean up code and implement minor fixes
Thanks, clang-tidy

+++ include/athenosum/beatmap.h
@@ -56,8 +56,9 @@ size_t parse_beatmap(char *file, struct osu_hitpoint **points,
  * Parses a total of count hitpoints from **points into **actions.
  * Returns the number of actions parsed and stored, which should be count * 2.
  */
-int parse_hitpoints(size_t count, size_t columns, struct osu_hitpoint **points,
-		    struct osu_action **actions);
+size_t parse_hitpoints(size_t count, size_t columns,
+		       struct osu_hitpoint **points,
+		       struct osu_action **actions);
 
 /**
  * Sort the array of actions given through **actions by time.
